refactor: extract composite range scan consumer
